Linux: strip out bundled JRE.
Require the system to provide a JRE for running Wiring IDE on
Linux. It's too much of a headache to bundle the JRE on Linux, so drop
it. This does not affect other operating systems.

Since librxtxSerial.so was previously placed in a default search
location in the bundled JRE, we now need to prepare the environment to
find it when running the wiring shell script. Do this by hacking
LD_LIBRARY_PATH. Perhaps editing java.library.path will also be
useful, but this seems to work.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>