build: Clean up init tasks.
Rename the top-level "-init" task to "init"; apparently this is a
target people expect to see, so we might as well give it to them.

Make init depend on platform.-init, so we can specify a single
dependency on "init" instead of on both init and platform.-init, which
is a mistake. (This is fine since the mkdir task guarantees parents
are created as necessary.)

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>